move pdf finder helper to metadata
