+++ shared/pub/metadata.js
